Rework SpriteAnimator
