+++ Nez.Portable/ECS/Components/Renderables/Sprites/Animation/AnimationLoopMode.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Nez.Sprites
+﻿namespace Nez.Sprites
 {
 	public partial class SpriteAnimator
 	{
